@@ -1,9 +1,8 @@
 ﻿using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
 using TestServer.Hubs;
 
 namespace TestServer
@@ -20,21 +19,11 @@ namespace TestServer
     // This method gets called by the runtime. Use this method to add services to the container.
     public void ConfigureServices(IServiceCollection services)
     {
-      services.Configure<CookiePolicyOptions>(options =>
-      {
-        // This lambda determines whether user consent for non-essential cookies is needed for a given request.
-        options.CheckConsentNeeded = context => true;
-        options.MinimumSameSitePolicy = SameSiteMode.None;
-      });
-
-
-      services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_1);
-
       services.AddSignalR();
     }
 
     // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-    public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+    public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
     {
       if (env.IsDevelopment())
       {
@@ -50,11 +39,14 @@ namespace TestServer
       app.UseDefaultFiles();
       app.UseStaticFiles();
       app.UseCookiePolicy();
-      app.UseSignalR(routes =>
+
+      app.UseRouting();
+
+      app.UseAuthorization();
+      app.UseEndpoints(endpoints =>
       {
-        routes.MapHub<IntegrationTestHub>("/IntegrationTestHub");
+        endpoints.MapHub<IntegrationTestHub>("/IntegrationTestHub");
       });
-      app.UseMvc();
     }
   }
 }
