@@ -2,7 +2,7 @@ import 'package:chatclient/app.dart';
 import 'package:flutter/material.dart';
 
 
-const kChatServerUrl = "http://192.168.10.50:51002";
+const kChatServerUrl = "http://192.168.0.24:5000";
 
 void main() {
   runApp(App());
