@@ -5,11 +5,13 @@ import 'http_connection.dart';
 import 'http_connection_options.dart';
 import 'hub_connection.dart';
 import 'ihub_protocol.dart';
+import 'iretry_policy.dart';
 import 'itransport.dart';
 import 'json_hub_protocol.dart';
 import 'utils.dart';
+import 'default_reconnect_policy.dart';
 
-/// A builder for configuring {@link @aspnet/signalr.HubConnection} instances.
+/// A builder for configuring {@link @microsoft/signalr.HubConnection} instances.
 class HubConnectionBuilder {
   // Properties
 
@@ -21,6 +23,10 @@ class HubConnectionBuilder {
 
   Logger _logger;
 
+  /// If defined, this indicates the client should automatically attempt to reconnect if the connection is lost. */
+  ///
+  IRetryPolicy _reconnectPolicy;
+
   /// Configures console logging for the HubConnection.
   ///
   /// logger: this logger with the already configured log level will be used.
@@ -31,7 +37,7 @@ class HubConnectionBuilder {
     return this;
   }
 
-  /// Configures the {@link @aspnet/signalr.HubConnection} to use HTTP-based transports to connect to the specified URL.
+  /// Configures the {@link @microsoft/signalr.HubConnection} to use HTTP-based transports to connect to the specified URL.
   ///
   /// url: The URL the connection will use.
   /// options: An options object used to configure the connection.
@@ -39,17 +45,16 @@ class HubConnectionBuilder {
   /// Use either options or transportType.
   /// Returns the builder instance, for chaining.
   ///
-  HubConnectionBuilder withUrl(String url,
-      {HttpConnectionOptions options, HttpTransportType transportTyp}) {
+  HubConnectionBuilder withUrl(String url, {HttpConnectionOptions options, HttpTransportType transportType}) {
     assert(!isStringEmpty(url));
-    assert(!(options != null && transportTyp != null));
+    assert(!(options != null && transportType != null));
 
     _url = url;
 
     if (options != null) {
       _httpConnectionOptions = options;
     } else {
-      _httpConnectionOptions = HttpConnectionOptions(transport: transportTyp);
+      _httpConnectionOptions = HttpConnectionOptions(transport: transportType);
     }
 
     return this;
@@ -66,6 +71,20 @@ class HubConnectionBuilder {
     return this;
   }
 
+  HubConnectionBuilder withAutomaticReconnect({IRetryPolicy reconnectPolicy, List<int> retryDelays}) {
+    assert(_reconnectPolicy == null);
+
+    if (reconnectPolicy == null && retryDelays == null) {
+      _reconnectPolicy = DefaultRetryPolicy();
+    } else if (retryDelays != null) {
+      _reconnectPolicy = DefaultRetryPolicy(retryDelays: retryDelays);
+    } else {
+      _reconnectPolicy = reconnectPolicy;
+    }
+
+    return this;
+  }
+
   /// Creates a HubConnection from the configuration options specified in this builder.
   ///
   /// Returns the configured HubConnection.
@@ -73,8 +92,7 @@ class HubConnectionBuilder {
   HubConnection build() {
     // If httpConnectionOptions has a logger, use it. Otherwise, override it with the one
     // provided to configureLogger
-    final httpConnectionOptions =
-        _httpConnectionOptions ?? HttpConnectionOptions();
+    final httpConnectionOptions = _httpConnectionOptions ?? HttpConnectionOptions();
 
     // Now create the connection
     if (isStringEmpty(_url)) {
@@ -82,6 +100,6 @@ class HubConnectionBuilder {
           "The 'HubConnectionBuilder.withUrl' method must be called before building the connection.");
     }
     final connection = HttpConnection(_url, options: httpConnectionOptions);
-    return HubConnection(connection, _logger, _protocol ?? JsonHubProtocol());
+    return HubConnection.create(connection, _logger, _protocol ?? JsonHubProtocol(), reconnectPolicy: _reconnectPolicy);
   }
 }
