@@ -4,6 +4,8 @@ import 'errors.dart';
 
 /// Specifies a specific HTTP transport type.
 enum HttpTransportType {
+  /// Specified no transport preference. */
+  None, // = 0,
   /// Specifies the WebSockets transport. */
   WebSockets, // = 1,
   /// Specifies the Server-Sent Events transport. */
@@ -11,9 +13,10 @@ enum HttpTransportType {
   /// Specifies the Long Polling transport. */
   LongPolling, // = 4,
 }
+
 HttpTransportType httpTransportTypeFromString(String value) {
-  if (value == null) {
-    return null;
+  if (value == null || value == "") {
+    return HttpTransportType.None;
   }
 
   value = value.toUpperCase();
@@ -34,6 +37,8 @@ HttpTransportType httpTransportTypeFromString(String value) {
 
 /// Specifies the transfer format for a connection.
 enum TransferFormat {
+  /// TransferFormat is not defined.
+  Undefined, // = 0,
   /// Specifies that only text data will be transmitted over the connection.
   Text, // = 1,
   /// Specifies that binary data will be transmitted over the connection.
@@ -41,8 +46,8 @@ enum TransferFormat {
 }
 
 TransferFormat getTransferFormatFromString(String value) {
-  if (value == null) {
-    return null;
+  if (value == null || value == "") {
+    return TransferFormat.Undefined;
   }
 
   value = value.toUpperCase();
@@ -63,7 +68,7 @@ TransferFormat getTransferFormatFromString(String value) {
 typedef OnReceive = void Function(Object data);
 
 ///
-typedef OnClose = void Function(Exception error);
+typedef OnClose = void Function({Exception error});
 
 typedef AccessTokenFactory = Future<String> Function();
 
@@ -73,7 +78,7 @@ abstract class ITransport {
 
   /// data: the content. Either a string (json) or Uint8List (binary)
   Future<void> send(Object data);
-  Future<void> stop(Error error);
+  Future<void> stop({Error error});
   OnReceive onReceive;
   OnClose onClose;
 }
