@@ -2,7 +2,7 @@ import 'package:client/app.dart';
 import 'package:flutter/material.dart';
 
 
-const kServerUrl = "http://192.168.2.107:51001";
+const kServerUrl = "http://192.168.0.24:5000";
 
 void main() {
   runApp(App());
